+++ GameFrameWork/AbstractPlayer.cs
@@ -18,25 +18,6 @@ namespace GameFrameWork
         }
     }
 
-    public abstract class AbstractHumanPlayer : AbstractPlayer
-    {
-        public AbstractHumanPlayer(string name, PlayerType type, object moveSymbol) : base(name, type, moveSymbol) { }
-        
-        // Human player -  method for selecting moves
-        public abstract object SelectMove(AbstractBoard board);
-    }
-
-    public abstract class AbstractComputerPlayer : AbstractPlayer
-    {
-        public AbstractComputerPlayer(string name, PlayerType type, object moveSymbol) : base(name, type, moveSymbol) { }
-
-        // For computer player - find winning moves
-        public abstract object FindWinningMove(AbstractBoard board);
-
-        // For computer player - select random move when no winning move
-        public abstract object SelectRandomMove();
-    }
-
     public enum PlayerType
     {
         Human,
